+++ client/Connection.py
@@ -1,27 +1,37 @@
 import socket
+import configparser
 
 class Connection:
 
-    def __init__(self, responses):
-        print("Connect")
-        self.socket = socket.socket()
-        self.socket.connect(('localhost', 12345))
-
-        self.responses = responses
+    def __init__(self, socket):
+        self.socket = socket
 
     def sendRequest(self, msg):
         self.socket.send(msg)
-        self.getResponse()
 
     def getResponse(self):
         buffer = ""
         BUFFERSIZE = 1024
+        self.socket.settimeout(5)
         while True:
-            msg = self.socket.recv(BUFFERSIZE)
+            print("read Socket!")
+            try:
+                msg = self.socket.recv(BUFFERSIZE)
+            except socket.timeout:
+                break
             if len(msg) == BUFFERSIZE:
                 buffer += msg.decode()
             else:
                 buffer += msg.decode()
                 break
 
-        self.responses.put(buffer)
+        return buffer
+
+
+def createSocket():
+    CONFIG = configparser.ConfigParser()
+    CONFIG.read('../config.py')
+
+    s = socket.socket()
+    s.connect((CONFIG.get('all', 'HOST'), CONFIG.getint('all', 'PORT')))
+    return s
