+++ config.py
@@ -1,0 +1,3 @@
+[all]
+HOST=localhost
+PORT=12335
